pyproject.toml
[project]
name = "gemini-gitlab-workflow"
version = "0.1.0"
authors = [{ name = "atistrcsn", email = "developer@hobbesmedia.hu" }]
description = "A comprehensive workflow tool for synchronizing GitLab issues to local markdown files, generating AI-assisted feature plans, and uploading them back to GitLab."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
keywords = ["GitLab", "AI", "Workflow", "Automation", "Gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-gitlab",
    "pyyaml",
    "typer",
    "python-dotenv",
    "networkx",
    "rich",
    "google-generativeai",
    "pydantic>=2.12.4",
    "ruamel-yaml>=0.18.16",
]

[project.scripts]
ggw = "gemini_gitlab_workflow.cli:app"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "pytest-cov>=7.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-mock",
    "pytest-cov>=7.0.0",
    "pyinstaller",
]
